+++ NEWS.md
@@ -1,3 +1,16 @@
+# Mhorseshoe 0.1.2
+
+This is a very small release focused on fixing issues seen in 
+approx_horseshoe's examples in 0.1.0 and the submitted 0.1.1 version.
+
+* Fix a problem that occurred in the examples of approx_horseshoe. This 
+problem occurs when the size of the data is not large enough and the threshold 
+argument is set too small when auto.threshold=FALSE.
+
+* Add code to the approx_horseshoe function: If a problem occurs because the 
+threshold argument is set too small, a warning is output and the result is 
+returned.
+
 # Mhorseshoe 0.1.1
 
 * Fixed bug in examples of approx_horseshoe